Implemented Reversed Indexing
Added the ability to reference cells relative to the bottom of the table.

I ran into an issue where I wanted to have a total and total after tax listed at the bottom of a table, but found no way to do this. I saw another user had a similar issue and submitted a feature request so I decided to give it a go! (https://github.com/tgrosinger/advanced-tables-obsidian/issues/334)

This works through the first '<' and last '>' argument combined with relative indexing ('+n' or '-n')

ex: @>-1 references the 2nd to last row and $>-2 references the 3rd to last column.

This also works with 2nd to first row and column: @<+2 and $<+1

I added test cases for these and functionality seems correct, but I'm not 100% sure that I tested all the scenarios.

Updated the formulas.md.

I slightly changed the functionality of the row and column classes. The index now starts at 0 instead of -1 and the switch case now adds the appropriate values to index. This allows index be changed relatively then add the absolute value.

I had to change .eslintrc.js to get it to run for me as I am on a windows machine and it probably needs reverted. Sorry if this (or anything else) isn't standard practice, this is my first contribution to a public GitHub.